Moved most of stmgr code into lib.
